split volunteer.cpp into diffrent files, 1 for each class
